Добавляет отрисовку фото
