+++ js/main.js
@@ -1,3 +1,3 @@
-import {getPhotos} from './rendering.js'; // модуль с функцией для генерации фотокарточек
-
+import {getPhotos} from './createphotos.js'; // модуль с функцией для генерации фотокарточек
+import {} from 'rendering.js';
 getPhotos();
